+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 # Change Log
 
-## [0.2.98] - 2018-09-17
+## [0.2.99] - 2018-09-19
+### Added
+- The `back button` and `back button label` modifiers.
+- The `show login` directive in interview `metadata`.
+
+## [0.2.98] - 2018-09-18
 ### Fixed
 - Error in Word manifest XML file.
 
-## [0.2.97] - 2018-09-17
+## [0.2.97] - 2018-09-18
 ### Added
 - Playground "Variables, etc." for the Word task pane.
 ### Changed